Only set parentName or parentVersion when both are set
For new project versions both values are required in order to resolve
the parent. The parentName defaults to ${project.parent.name}, but
parentVersion defaults to null. This combination would result in a 404
response from Dependency Track when uploading a new (child) project
version.

As parentName is also always set it also prevents a project from not
being registered as a child project. Unsetting the default value is not
possible from the pom.xml, only via a system property.

See issue #441