Changed BOM uploading to using the POST API by default
When needed users can still use the PUT API to upload by setting the
configuration flag `uploadWithPut`. Primary reason for switching to the
POST API is because it plays nicer with WebApplication Firewalls in case
of large request payloads.

Note that the POST and PUT APIs have subtle differences in the request
parameters.

Further more, the BOM content is no longer kept in memory during upload,
instead an InputStream is passed on to the different request handlers.
This should also have a positive effect when processing huge SBOMs.