Upload BOM using the POST API (#455)
* Updated commons-logging to allow HTTP debug logging in maven

The commons-logging pulled in by the Apache HttpClient used by Unirest
does not support slf4j, which is used by Maven. Due to this you cannot
easily enable HttpClient's debug logging.

* Changed BOM uploading to using the POST API by default

When needed users can still use the PUT API to upload by setting the
configuration flag `uploadWithPut`. Primary reason for switching to the
POST API is because it plays nicer with WebApplication Firewalls in case
of large request payloads.

Note that the POST and PUT APIs have subtle differences in the request
parameters.

Further more, the BOM content is no longer kept in memory during upload,
instead an InputStream is passed on to the different request handlers.
This should also have a positive effect when processing huge SBOMs.

* Do not use the standalone Unirest build

* Pass the bom File if a file reference

This results in the Content-Length being set in the request.

* Set HTTP user agent to plugin's name and version

* Changed default value for uploadWithPut to true

The default configuration will have the same behavior as previous plugin
versions. A future major release might change the default behavior to
using the POST API.

---------

Co-authored-by: Paul McKeown <pmckeown@gmail.com>