Updated mojo tests to use the new maven plugin testing harness
It contains a few workarounds to deal with shortcomings in version 3.4.0

The maven project is a rather shallow mock where various elements are
not populated. e.g. `getBuild()` returns null.
The referenced pom is also not consulted for mojo property expression
evaluation. So the default projectName and projectVersion remain null.