Junit 5 upgrade (#459) (#460)
* Junit 5 upgrade (#459)

* Migration to JUnit5

* Apply junit5 best practices

* Update test dependencies

* Improve test quality

- Parameterized the BomParserTest
- Removed unused variables
- Resolved generics
- Resolved some possible NPE warnings

* Updated mojo tests to use the new maven plugin testing harness

It contains a few workarounds to deal with shortcomings in version 3.4.0

The maven project is a rather shallow mock where various elements are
not populated. e.g. `getBuild()` returns null.
The referenced pom is also not consulted for mojo property expression
evaluation. So the default projectName and projectVersion remain null.

* Run verify phases in maven build job

Execute unit tests and spotless

* Use strict mock verification

* Renamed test

* Tabs to spaces

---------

Co-authored-by: Michiel Hendriks <elmuerte@users.noreply.github.com>
Co-authored-by: Paul McKeown <paul@midships.io>